Folding akd_mysql, integration_tests, and akd_test_tools into examples
