+++ Cargo.toml
@@ -4,10 +4,7 @@ members = [
     "akd_core",
     "akd",
     "akd_client",
-    "akd_mysql",
-    "akd_test_tools",
     "examples",
-    "integration_tests",
     "xtask",
 ]
