+++ akd/Cargo.toml
@@ -39,17 +39,17 @@ default = ["public_auditing", "parallel_vrf", "parallel_insert", "preload_histor
 akd_core = { path = "../akd_core", default-features = false, features = ["vrf"] }
 async-recursion = "1"
 async-trait = "0.1"
-dashmap = { version = "5" }
+dashmap = "5"
 hex = "0.4"
-log = { version = "0.4.8", features = ["kv_unstable"] }
-tokio = { version = "1.21", features = ["sync", "time", "rt"] }
+log = { version = "0.4", features = ["kv_unstable"] }
+tokio = { version = "1", features = ["sync", "time", "rt"] }
 
 ## Optional dependencies ##
 serde = { version = "1", features = ["derive"], optional = true }
 rand = { version = "0.8", optional = true }
 colored = { version = "2", optional = true }
 once_cell = { version = "1", optional = true }
-protobuf = { version = "3.2", optional = true }
+protobuf = { version = "3", optional = true }
 paste = { version = "1", optional = true }
 
 [dev-dependencies]
@@ -57,11 +57,11 @@ criterion = "0.5"
 serial_test = "2"
 proptest = "1.0"
 proptest-derive = "0.3"
-colored = { version = "2" }
-once_cell = { version = "1" }
+colored = "2"
+once_cell = "1"
 ctor = "0.2"
 tokio-test = "0.4"
-tokio = { version = "1.21", features = ["rt", "sync", "time", "macros"] }
+tokio = { version = "1", features = ["rt", "sync", "time", "macros"] }
 mockall = "0.11"
 futures = "0.3"
 itertools = "0.11"
