+++ akd_core/Cargo.toml
@@ -12,8 +12,8 @@ readme = "../README.md"
 build = "src/build.rs"
 
 [build-dependencies]
-protobuf-codegen = "3.2"
-protobuf-parse = "3.2"
+protobuf-codegen = "3"
+protobuf-parse = "3"
 
 [features]
 # Disable all STD for the crate
@@ -43,16 +43,16 @@ hex = "0.4"
 
 ## Optional dependencies ##
 blake3 = { version = "1.3", optional = true, default-features = false }
-protobuf = { version = "3.2", optional = true }
+protobuf = { version = "3", optional = true }
 rand = { version = "0.8", optional = true }
 serde = { version = "1", features = ["derive"], optional = true }
 serde_bytes = { version = "0.11", optional = true }
-tokio = { version = "1.21", features = ["rt"], optional = true }
+tokio = { version = "1", features = ["rt"], optional = true }
 paste = { version = "1", optional = true }
 
 [dev-dependencies]
 bincode = "1"
-proptest = "1.0"
+proptest = "1"
 proptest-derive = "0.3"
 rand = "0.8"
 serde = { version = "1", features = ["derive"] }
