--- akd_mysql/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "akd_mysql"
-version = "0.9.0-pre.1"
-authors = ["Harjasleen Malvai <hmalvai@fb.com>", "Kevin Lewi <klewi@fb.com>", "Sean Lawlor <seanlawlor@fb.com>"]
-description = "A MySQL storage layer implementation for an auditable key directory (AKD)"
-license = "MIT OR Apache-2.0"
-edition = "2018"
-keywords = ["key-transparency", "akd", "mysql", "akd-mysql"]
-repository = "https://github.com/facebook/akd"
-
-[features]
-# Collect runtime metrics on db access calls + timing
-runtime_metrics = ["akd/runtime_metrics"]
-whatsapp_v1 = ["akd/whatsapp_v1"]
-experimental = ["akd/experimental"]
-default = ["experimental"]
-
-[dependencies]
-rand = "0.8"
-serde = { version = "1", features = ["derive"] }
-async-trait = "0.1"
-tokio = { version = "1.21", features = ["full"] }
-async-recursion = "1"
-mysql_async = "0.32"
-mysql_common = "0.30"
-log = { version = "0.4.8", features = ["kv_unstable"] }
-akd = { path = "../akd", features = ["serde_serialization"], default-features = false }
-
-[dev-dependencies]
-criterion = "0.5"
-serial_test = "2"
-akd = { path = "../akd", features = ["experimental", "public_tests"], default-features = false }
