--- akd_test_tools/Cargo.toml
@@ -1,33 +1,0 @@
-[package]
-name = "akd_test_tools"
-version = "0.9.0-pre.1"
-authors = ["Evan Au <evanau@fb.com>", "Sean Lawlor <seanlawlor@fb.com>"]
-description = "Test utilities and tooling"
-license = "MIT OR Apache-2.0"
-edition = "2021"
-keywords = ["key-transparency", "akd"]
-repository = "https://github.com/facebook/akd"
-readme = "../README.md"
-
-[features]
-whatsapp_v1 = ["akd/whatsapp_v1"]
-default = ["whatsapp_v1"]
-
-[dependencies]
-clap = { version = "4", features = ["derive"] }
-log = { version = "0.4.8", features = ["kv_unstable"] }
-tokio = { version = "1.21", features = ["full"] }
-regex = "1.5"
-rand = "0.8"
-serde_yaml = "0.9"
-serde = "1.0"
-async-trait = "0.1"
-thread-id = "4"
-
-akd = { path = "../akd", features = ["serde_serialization"] }
-
-[dev-dependencies]
-assert_fs = "1"
-paste = "1"
-
-akd = { path = "../akd", features = ["public_tests", "serde_serialization"] }
