+++ examples/Cargo.toml
@@ -16,13 +16,15 @@ doc = false
 [dependencies]
 anyhow = "1"
 async-trait = "0.1"
-colored = "2.0.0"
+colored = "2"
 clap = { version = "4", features = ["derive"] }
 dialoguer = "0.10"
 hex = "0.4"
 indicatif = "0.17"
-log = { version = "0.4.8", features = ["kv_unstable"] }
+log = { version = "0.4", features = ["kv_unstable"] }
 multi_log = "0.1"
+mysql_async = "0.32"
+mysql_common = "0.30"
 once_cell = "1"
 rand = "0.8"
 serde = { version = "1.0", features = ["derive"] }
@@ -31,9 +33,12 @@ thread-id = "4"
 tokio = { version = "1.21", features = ["full"] }
 xml-rs = "0.8"
 reqwest = "0.11"
+regex = "1"
+serde_yaml = "0.9"
 
 akd = { path = "../akd", features = ["public_tests", "public_auditing", "whatsapp_v1", "experimental"] }
-akd_mysql = { path = "../akd_mysql" }
 
 [dev-dependencies]
-serial_test = "*"
+serial_test = "2"
+assert_fs = "1"
+paste = "1"
