+++ examples/src/fixture_generator/examples/example_tests.rs
