+++ examples/src/fixture_generator/generator.rs
@@ -17,7 +17,6 @@ use akd::directory::Directory;
 use akd::storage::types::DbRecord;
 use akd::storage::{StorageManager, StorageUtil};
 use akd::{AkdLabel, AkdValue, NamedConfiguration};
-use clap::Parser;
 use rand::rngs::StdRng;
 use rand::Rng;
 use rand::SeedableRng;
@@ -56,15 +55,13 @@ const HEADER_COMMENT: &str = concat!(
     "@",
     "generated This file was automatically generated by \n\
     the fixture generator tool with the following command:\n\n\
-    cargo run -- \\"
+    cargo run -p examples -- fixture-generator \\"
 );
 const METADATA_COMMENT: &str = "Metadata";
 const STATE_COMMENT: &str = "State - Epoch";
 const DELTA_COMMENT: &str = "Delta - Epoch";
 
-pub async fn run() {
-    let args = Args::parse();
-
+pub async fn run(args: Args) {
     // NOTE(new_config): Add new configurations here
     generate::<akd::WhatsAppV1Configuration>(&args).await;
     generate::<akd::ExperimentalConfiguration>(&args).await;
