+++ examples/src/fixture_generator/parser.rs
