+++ examples/src/fixture_generator/reader/mod.rs
