+++ examples/src/fixture_generator/reader/tests.rs
