+++ examples/src/fixture_generator/writer/mod.rs
