+++ examples/src/fixture_generator/writer/yaml.rs
