+++ examples/src/mysql_demo/mysql_storables.rs
