+++ examples/src/mysql_demo/tests/mysql_tests.rs
@@ -5,15 +5,18 @@
 // License, Version 2.0 found in the LICENSE-APACHE file in the root directory
 // of this source tree. You may select, at your option, one of the above-listed licenses.
 
+use crate::mysql_demo::mysql::AsyncMySqlDatabase;
+use crate::mysql_demo::tests::test_util::{
+    directory_test_suite, log_init, test_lookups as test_lookups_util,
+};
 use crate::test_config_serial;
 use akd::storage::StorageManager;
 use akd::{ecvrf::HardCodedAkdVRF, Configuration};
-use akd_mysql::mysql::*;
 use log::{error, info, warn};
 
 test_config_serial!(test_directory_operations);
 async fn test_directory_operations<TC: Configuration>() {
-    crate::test_util::log_init(log::Level::Info);
+    log_init(log::Level::Info);
 
     info!("\n\n******** Starting MySQL Directory Operations Integration Test ********\n\n");
 
@@ -49,12 +52,7 @@ async fn test_directory_operations<TC: Configuration>() {
 
         let vrf = HardCodedAkdVRF {};
         let storage_manager = StorageManager::new_no_cache(mysql_db.clone());
-        akd_test_tools::test_suites::directory_test_suite::<TC, _, HardCodedAkdVRF>(
-            &storage_manager,
-            50,
-            &vrf,
-        )
-        .await;
+        directory_test_suite::<TC, _, HardCodedAkdVRF>(&storage_manager, 50, &vrf).await;
 
         storage_manager.log_metrics(log::Level::Trace).await;
 
@@ -75,7 +73,7 @@ async fn test_directory_operations<TC: Configuration>() {
 
 test_config_serial!(test_directory_operations_with_caching);
 async fn test_directory_operations_with_caching<TC: Configuration>() {
-    crate::test_util::log_init(log::Level::Info);
+    log_init(log::Level::Info);
 
     info!("\n\n******** Starting MySQL Directory Operations (w/caching) Integration Test ********\n\n");
 
@@ -111,12 +109,7 @@ async fn test_directory_operations_with_caching<TC: Configuration>() {
 
         let vrf = HardCodedAkdVRF {};
         let storage_manager = StorageManager::new(mysql_db.clone(), None, None, None);
-        akd_test_tools::test_suites::directory_test_suite::<TC, _, HardCodedAkdVRF>(
-            &storage_manager,
-            50,
-            &vrf,
-        )
-        .await;
+        directory_test_suite::<TC, _, HardCodedAkdVRF>(&storage_manager, 50, &vrf).await;
 
         storage_manager.log_metrics(log::Level::Trace).await;
 
@@ -137,7 +130,7 @@ async fn test_directory_operations_with_caching<TC: Configuration>() {
 
 test_config_serial!(test_lookups);
 async fn test_lookups<TC: Configuration>() {
-    crate::test_util::log_init(log::Level::Info);
+    log_init(log::Level::Info);
 
     info!("\n\n******** Starting MySQL Lookup Tests ********\n\n");
 
@@ -174,14 +167,7 @@ async fn test_lookups<TC: Configuration>() {
         let vrf = HardCodedAkdVRF {};
         let storage_manager = StorageManager::new(mysql_db, None, None, None);
 
-        crate::test_util::test_lookups::<TC, _, HardCodedAkdVRF>(
-            &storage_manager,
-            &vrf,
-            50,
-            5,
-            100,
-        )
-        .await;
+        test_lookups_util::<TC, _, HardCodedAkdVRF>(&storage_manager, &vrf, 50, 5, 100).await;
 
         // clean the test infra
         if let Err(mysql_async::Error::Server(error)) = storage_manager.get_db().drop_tables().await
