--- integration_tests/Cargo.toml
@@ -1,26 +1,0 @@
-[package]
-name = "akd_integration_tests"
-version = "0.0.0"
-authors = ["Harjasleen Malvai <hmalvai@fb.com>", "Kevin Lewi <klewi@fb.com>", "Sean Lawlor <seanlawlor@fb.com>"]
-edition = "2021"
-publish = false
-
-[features]
-
-[dependencies]
-akd = { path = "../akd", features = ["public_tests", "serde_serialization", "runtime_metrics", "whatsapp_v1", "experimental"] }
-
-[dev-dependencies]
-log = { version = "0.4.8", features = ["kv_unstable"] }
-tokio = { version = "1.21", features = ["full"] }
-serial_test = "*"
-mysql_async = "0.32"
-rand = "0.8"
-once_cell = "1"
-thread-id = "4"
-multi_log = "0.1"
-hex = "0.4.3"
-paste = "1"
-
-akd_mysql = { path = "../akd_mysql", features = ["runtime_metrics"] }
-akd_test_tools = { path = "../akd_test_tools" }
