--- integration_tests/src/lib.rs
@@ -1,14 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-//
-// This source code is dual-licensed under either the MIT license found in the
-// LICENSE-MIT file in the root directory of this source tree or the Apache
-// License, Version 2.0 found in the LICENSE-APACHE file in the root directory
-// of this source tree. You may select, at your option, one of the above-listed licenses.
-
-#[cfg(test)]
-mod mysql_tests;
-#[cfg(test)]
-mod test_util;
-
-#[cfg(test)]
-mod memory_tests;
